% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfc_lignin_test_dups}
\alias{cfc_lignin_test_dups}
\title{Lignin data table from Plant foliar traits (DP1.10026.001)}
\format{
A data frame with 26 rows and 25 columns
\describe{
  The variable names, descriptions, and units can be found in the cfc_lignin_variables table
}
}
\source{
\url{https://data.neonscience.org/api/v0/products/DP1.10026.001}
}
\usage{
cfc_lignin_test_dups
}
\description{
An example set of NEON observational data, containing duplicate records. NOT APPROPRIATE FOR ANALYTICAL USE. Contains foliar lignin data from Moab and Toolik in 2017, with artificial duplicates introduced to demonstrate the removeDups() function.
}
\keyword{datasets}
