% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupProcess.R
\name{dupProcess}
\alias{dupProcess}
\title{Helper function to remove duplicates from a data table; assumes input data are a duplicated set.}
\usage{
dupProcess(data, data.dup, table)
}
\arguments{
\item{data}{A data frame containing original duplicated data. [data frame]}

\item{data.dup}{A data frame containing lowercase conversion of the duplicated data. [character]}

\item{table}{The table name for the input data frame}
}
\value{
A modified data frame with resolveable duplicates removed and a flag field added and populated.
}
\description{
Helper function to carry out duplicate removal on a data table of duplicates.
}
\details{
Helper function to carry out the flagging and removal for removeDups().
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
