% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleParents.R
\name{getSampleParents}
\alias{getSampleParents}
\title{Find all ancestors of a given sample.}
\usage{
getSampleParents(sampleUuid, token = NA_character_)
}
\arguments{
\item{sampleUuid}{A NEON sample UUID. [character]}

\item{token}{User specific API token (generated within neon.datascience user accounts). Optional. [character]}
}
\value{
A table of sample identifiers, their classes, and their parent samples.
}
\description{
Find all samples in the upstream sample tree of a given sample. Used in getSampleTree().
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
