% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleTree.R
\name{getSampleTree}
\alias{getSampleTree}
\title{Find all relatives (parents, children, and outward) of a given sample.}
\usage{
getSampleTree(
  sampleNode,
  idType = "tag",
  sampleClass = NA_character_,
  token = NA_character_
)
}
\arguments{
\item{sampleNode}{A NEON sample identifier. [character]}

\item{idType}{Is sampleNode a tag, barcode, or guid? Defaults to tag. [character]}

\item{sampleClass}{The NEON sampleClass of sampleNode. Required if sampleNode is a tag and there are multiple valid classes. [character]}

\item{token}{User specific API token (generated within neon.datascience user accounts). Optional. [character]}
}
\value{
A table of sample identifiers, their classes, and their parent samples.
}
\description{
Find all samples in the sample tree of a given sample.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
