% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idSampleParents.R
\name{idSampleParents}
\alias{idSampleParents}
\title{Reformat sample data to indicate the parents of a focal sample.}
\usage{
idSampleParents(sampleUuid, token = NA_character_)
}
\arguments{
\item{sampleUuid}{A NEON sample UUID. [character]}

\item{token}{User specific API token (generated within neon.datascience user accounts). Optional. [character]}
}
\value{
A table of sample identifiers, their classes, and their parent samples.
}
\description{
Reformat table of sample identifiers to include parent sample identifiers. Used in getSampleTree().
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org}
}
